Save point for Room caching code
App crashes

+++ shared/src/commonMain/kotlin/com/vickbt/shared/utils/SharedExtensions.kt
@@ -52,6 +52,6 @@ fun Double.getRating(): String {
     return "$before.$after"
 }
 
-fun Long.toBoolean(): Boolean {
-    return this != 0L
+fun Int.toBoolean(): Boolean {
+    return this != 0
 }
